CIRC-473: Split Effective location test into API and unit.
